requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"
apache-airflow-providers-ssh
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
apache-airflow-providers-mysql
requests
apache-airflow-providers-jira


# Numeric utilities
numpy
scipy
scikit-learn
statsmodels
prophet
pandas

# COSMO
db_dtypes
google-cloud-bigquery
